Bluetooth: mt7921s: Enable SCO over I2S
The driver has to issue the specific command to enable Bluetooth SCO over
the I2S/PCM interface on mt7921s, that is supported since the firmware
with version 20211222191101 was added, and the patch would not cause any
harm even when the old firmware is applied.

The SCO profile with the patch was tested by setting up a VOIP application,
connected to HFP device, checked telephony function can work normally.

Co-developed-by: Sean Wang <sean.wang@mediatek.com>
Signed-off-by: Sean Wang <sean.wang@mediatek.com>
Signed-off-by: Mark Chen <mark-yw.chen@mediatek.com>
---
v2: refine git message and fix typo
v3:
   1. free skb after calling  __hci_cmd_sync
   2. make bt_awake as const struct btmtk_sco
v4:
   1. update git message
   2. drop a few redundant error messages
v5:
   1. make reg_write as const struct reg_write_cmd
   2. cleanup the error message in btsdio_mtk_reg_write
   3. rename btsdio_mtk_sco_setting of the prefix btmtksdio
   4. rename btsdio_mtk_reg_write of the prefix btmtksdio