mt76: sdio: honor the largest Tx buffer the hardware can support
We should take it to account the actual the host and the device MMC
capability to determine what the proper xmit_buf_size can be.

Both MT7921S and MT7663 can support up to Tx FIFO size of 0x3fe00 which
means the device can accept 511 blocks of block size 512 in a row from
the host. So if the driver aggregates the frames as many as possible the
device is able to support, we can merge multiple MMC requests into a
single one to save the overhead of the handling and synchronizing in
additional MMC requests, reduce the SDIO lock contention with the Bluetooth
concurrent traffic and eventually to have the higher bus utilization with
less idle cycle.

With the patch, it is helpful for WiFi to have the steady throughput
performance even while the Bluetooth is running in the background.

Co-developed-by: Sean Wang <sean.wang@mediatek.com>
Signed-off-by: Sean Wang <sean.wang@mediatek.com>
Signed-off-by: Deren Wu <deren.wu@mediatek.com>
Change-Id: I32c69e0fb99ccd09cb6006ad1504d588150a3aa0