mt76: mt7921: introduce chanctx support
The fimrware can have the capability to manage the channel context
scheduling on multiple roles running on the device including Station,
AP and P2P GC/GO mode (will be extended based on the patcheset ) to help
users sharing the network to the others in the single device.

The firmware is able to support the channel chanctx up to 2 interface
simultaneously running on the different channels.

Another thing to be noted is before driver is going to send out the managemnt
frames, the driver have to get the priviledge from the firmware to ocuppy the
channel context until the handshake is completed and then get the privilege
back to the firmware.

Signed-off-by: Deren Wu <deren.wu@mediatek.com>
Signed-off-by: Sean Wang <sean.wang@mediatek.com>