mt76: mt7921e: fix race issue between reset and suspend/resume
It is unexpected that the reset work is running simultaneously with
the suspend/resume context and it is possible reset work is still
alive after mt7921 is suspended if we don't fix the race issue.

Thus, the suspend should be synchronized until the reset is completed
at the beginning and ignore the subsequent requests.

In case there is an error that happens during either suspend or resume
handler, we will schedule a reset task to recover the error before
returning the error code of mt7921 suspend/resume to ensure race
free between reset and suspend/resume.

Fixes: 0c1ce9884607 ("mt76: mt7921: add wifi reset support")
Co-developed-by: YN Chen <YN.Chen@mediatek.com>
Signed-off-by: YN Chen <YN.Chen@mediatek.com>
Signed-off-by: Sean Wang <sean.wang@mediatek.com>